Move is_any_handler_alive as a Server struct method
